+++ src/hardware.c
@@ -1,26 +1,33 @@
-#include "hardware.h"
-
 // Memory && IO operations.
 
-void memCheck(void* pointer)
+#include "hardware.h"
+
+void pointerCheck(void *ptr, char *errmsg) // Eg. malloc or FILE*.
 {
-	if(pointer == NULL)
+	if(ptr == NULL)
 	{
-		fputs("Pointer error!\n", stderr);
+		fprintf(stderr, "%s\n", errmsg);
 		exit(1);
 	}
+}
+
+void sigHandler(int nothing)
+{
+	if(nothing == 0) {}
 }
 
 char unixGetch(void) // https://stackoverflow.com/questions/12710582/
 {
 	char key;
 	struct termios oldt, newt;
-	tcgetattr(STDIN_FILENO, &oldt);
+
+	tcgetattr(STDIN_FILENO, &oldt); // Get terminal attribiutes.
 	newt = oldt;
 	newt.c_lflag &= ~(ICANON | ECHO);
-	tcsetattr(STDIN_FILENO, TCSANOW, &newt);
+	tcsetattr(STDIN_FILENO, TCSANOW, &newt); // Set terminal attribiutes.
+
 	key = getchar();
-	tcsetattr(STDIN_FILENO, TCSANOW, &oldt);
+	tcsetattr(STDIN_FILENO, TCSANOW, &oldt); // Set terminal attribiutes.
 	return key;
 }