+++ src/hardware.h
@@ -7,16 +7,8 @@
 #include <unistd.h>
 #include <termios.h>
 
-// Some special ASCII decimal codes.
-#define ENTER 10
-#define CTRL_X 24
-#define ARROW_UP 65
-#define ARROW_DOWN 66
-#define ARROW_RIGHT 67
-#define ARROW_LEFT 68
-#define BACKSPACE 127
-
-void memErrorCheck(void* pointer);
+void pointerCheck(void *ptr, char *errmsg);
+void sigHandler(int nothing);
 char unixGetch(void);
 
 #endif