+++ docs/CHANGELOG.md
@@ -1,12 +1,32 @@
-# Fiflo - changelog
+# fiflo - changelog
 
 ## Description
 Below version numbers contains more detailed descriptions of annotated Git
 tags. Dates are not alternative version names. They are just release dates.
 Changelog contains only new tangible features. To see what's new in source
-files use diff-like tools or "git log".
+files use git-show or git-log.
 
 ## Versions:
+### v1.0.0 (July 15, 2018):
+- Simplified API.
+- Less memory usage.
+- Proper text allocation.
+- Faster initialization.
+- Usable for the first time.
+- A lot of bugs removed.
+- UI redesigned.
+- 80 chars per line limit && auto newline.
+
+### v0.5.0-unstable (July 1, 2018):
+The last v0.y.z version and showup of the old program logic. Mostly insecure.
+Not merged with "master" so look for it with tags.
+- Multiple lines writing && rendering.
+- File input/output operations.
+- Man-page added.
+- Upper bar is back.
+- Absolute paths support.
+- OS check.
+
 ### v0.4.3 (June 10, 2018):
 - Info about min term size fix.